Move HAL implementation out of the src folder
