+++ hal/adafruit_nrf/bluetooth/elk_service.cpp
