+++ simulator/compile.cmake
@@ -103,17 +103,18 @@ set(SRC_SYSTEM_LOGIC
     ${LMBD_ROOT_DIR}/src/system/logic/sunset_timer.cpp
 )
 
+# The compile chain should have copied the correct HAL here
 set(SIMULATOR_HAL
-    ${LMBD_ROOT_DIR}/simulator/hal/bluetooth_mock.cpp
-    ${LMBD_ROOT_DIR}/simulator/hal/gpio_mock.cpp
-    ${LMBD_ROOT_DIR}/simulator/hal/i2c_mock.cpp
-    ${LMBD_ROOT_DIR}/simulator/hal/filesystem.cpp
-    ${LMBD_ROOT_DIR}/simulator/hal/pdm_handle_mock.cpp
-    ${LMBD_ROOT_DIR}/simulator/hal/serial_mock.cpp
-    ${LMBD_ROOT_DIR}/simulator/hal/queues_mock.cpp
-    ${LMBD_ROOT_DIR}/simulator/hal/registers_mock.cpp
-    ${LMBD_ROOT_DIR}/simulator/hal/threads_mock.cpp
-    ${LMBD_ROOT_DIR}/simulator/hal/time_mock.cpp
+    ${LMBD_ROOT_DIR}/src/generated/hal/bluetooth.cpp
+    ${LMBD_ROOT_DIR}/src/generated/hal/gpio.cpp
+    ${LMBD_ROOT_DIR}/src/generated/hal/i2c.cpp
+    ${LMBD_ROOT_DIR}/src/generated/hal/filesystem.cpp
+    ${LMBD_ROOT_DIR}/src/generated/hal/pdm_handle.cpp
+    ${LMBD_ROOT_DIR}/src/generated/hal/serial.cpp
+    ${LMBD_ROOT_DIR}/src/generated/hal/queues.cpp
+    ${LMBD_ROOT_DIR}/src/generated/hal/registers.cpp
+    ${LMBD_ROOT_DIR}/src/generated/hal/threads.cpp
+    ${LMBD_ROOT_DIR}/src/generated/hal/time.cpp
 )
 
 set(SIMULATOR_STATE
